Added option to only dump out database entries (without getting files in storage)

+++ scripts/firebase-dump/firebase-dump.py
@@ -11,6 +11,7 @@ def parse_args():
     parser.add_argument('serviceCredentials', help='path to app service credentials json file')
     parser.add_argument('-o', '--output', help='path to output directory')
     parser.add_argument('-c', '--clear', help='clear Firebase database and storage', action='store_true')
+    parser.add_argument('-d', '--database-only', help='only retrieve Firebase database (storage files will NOT be retrieved or deleted)', action='store_true')
     parser.add_argument('--clear-no-prompt', help='clear Firebase database and storage without a warning prompt', action='store_true')
 
     return parser.parse_args()
@@ -23,6 +24,7 @@ def read_json(path):
 args = parse_args()
 services_json_data = read_json(args.servicesJson)
 creds_json_path = args.serviceCredentials
+db_only = args.database_only
 clear_fb = args.clear or args.clear_no_prompt
 
 outdir = os.getcwd()
@@ -83,21 +85,22 @@ if uuids.each() != None:
 
 
 # Dump storage (and clear if specified)
-storage = firebase.storage()
-for file in storage.list_files():
-    filepath = file.name
-    tokens = filepath.split('/')
+if not db_only:
+    storage = firebase.storage()
+    for file in storage.list_files():
+        filepath = file.name
+        tokens = filepath.split('/')
 
-    if len(tokens) == 2:
-        uuid = tokens[0]
-        filename = tokens[1]
+        if len(tokens) == 2:
+            uuid = tokens[0]
+            filename = tokens[1]
 
-        out_subdir = os.path.join(outdir, uuid)
-        if not os.path.exists(out_subdir):
-            os.makedirs(out_subdir)
+            out_subdir = os.path.join(outdir, uuid)
+            if not os.path.exists(out_subdir):
+                os.makedirs(out_subdir)
 
-        file.download_to_filename(os.path.join(out_subdir, filename))
+            file.download_to_filename(os.path.join(out_subdir, filename))
 
-    if clear_fb:
-        file.delete()
+        if clear_fb:
+            file.delete()
